refactor: extract helpers
